Progress bar added and fixed some issues
