+++ .gitignore
@@ -1,7 +1,10 @@
 *.pyc
-venv/
+env/
 *.egg-info
 UI_Files/
 build/
 dist/
 .idea*
+pipgui/test/*
+.vscode
+
