--- pip_gui/Resource_Files/genreFile.json
@@ -1,1 +1,0 @@
-{"Foreign Function Interface": "#foreign-function-interface", "Interactive Interpreter": "#interactive-interpreter", "Hardware": "#hardware", "Text Processing": "#text-processing", "Tagging": "#tagging", "Data Analysis": "#data-analysis", "Resources": "#resources", "Distribution": "#distribution", "Computer Vision": "#computer-vision", "News Feed": "#news-feed", "Email": "#email", "Package Repositories": "#package-repositories", "Testing": "#testing", "RESTful API": "#restful-api", "Cryptography": "#cryptography", "Database": "#database", "Code Analysis and Linter": "#code-analysis-and-linter", "Web Frameworks": "#web-frameworks", "HTTP": "#http", "Web Crawling": "#web-crawling", "Miscellaneous": "#miscellaneous", "GUI": "#gui", "Network Virtualization": "#network-virtualization", "Files": "#files", "RPC Servers": "#rpc-servers", "WSGI Servers": "#wsgi-servers", "Implementations": "#implementations", "Services": "#services", "Database Drivers": "#database-drivers", "URL Manipulation": "#url-manipulation", "Documentation": "#documentation", "Machine Learning": "#machine-learning", "CMS": "#cms", "Anti-spam": "#anti-spam", "Package Management": "#package-management", "Date and Time": "#date-and-time", "MapReduce": "#mapreduce", "Search": "#search", "Command-line Tools": "#command-line-tools", "Podcasts": "#podcasts", "HTML Manipulation": "#html-manipulation", "Data Visualization": "#data-visualization", "Downloader": "#downloader", "Logging": "#logging", "Admin Panels": "#admin-panels", "Third-party APIs": "#third-party-apis", "Forms": "#forms", "Twitter": "#twitter", "Permissions": "#permissions", "WebSocket": "#websocket", "Audio": "#audio", "Static Site Generator": "#static-site-generator", "Compatibility": "#compatibility", "Data Validation": "#data-validation", "Caching": "#caching", "Networking": "#networking", "Contributing": "#contributing", "Concurrency and Parallelism": "#concurrency-and-parallelism", "Websites": "#websites", "Specific Formats Processing": "#specific-formats-processing", "Other Awesome Lists": "#other-awesome-lists", "DevOps Tools": "#devops-tools", "ORM": "#orm", "Configuration": "#configuration", "Editor Plugins and IDEs": "#editor-plugins-and-ides", "Functional Programming": "#functional-programming", "Web Content Extracting": "#web-content-extracting", "Authentication": "#authentication", "Template Engine": "#template-engine", "Job Scheduler": "#job-scheduler", "Algorithms and Design Patterns": "#algorithms-and-design-patterns", "Video": "#video", "Build Tools": "#build-tools", "Queue": "#queue", "Code Quality": "#code-quality", "Processes": "#processes", "Serialization": "#serialization", "Debugging Tools": "#debugging-tools", "Imagery": "#imagery", "Geolocation": "#geolocation", "Asset Management": "#asset-management", "Internationalization": "#internationalization", "E-commerce": "#e-commerce", "Game Development": "#game-development", "Deep Learning": "#deep-learning", "Serverless Frameworks": "#serverless-frameworks", "ChatOps Tools": "#chatops-tools", "Continuous Integration": "#continuous-integration", "Weekly": "#weekly", "Environment Management": "#environment-management", "Science": "#science", "Natural Language Processing": "#natural-language-processing"}
