--- pip_gui/Resource_Files/genreListFile.json
@@ -1,1 +1,0 @@
-["Admin Panels", "Algorithms and Design Patterns", "Anti-spam", "Asset Management", "Audio", "Authentication", "Build Tools", "Caching", "ChatOps Tools", "CMS", "Code Analysis and Linter", "Command-line Tools", "Compatibility", "Computer Vision", "Concurrency and Parallelism", "Configuration", "Cryptography", "Data Analysis", "Data Validation", "Data Visualization", "Database Drivers", "Database", "Date and Time", "Debugging Tools", "Deep Learning", "DevOps Tools", "Distribution", "Documentation", "Downloader", "E-commerce", "Editor Plugins and IDEs", "Email", "Environment Management", "Files", "Foreign Function Interface", "Forms", "Functional Programming", "Game Development", "Geolocation", "GUI", "Hardware", "HTML Manipulation", "HTTP", "Imagery", "Implementations", "Interactive Interpreter", "Internationalization", "Job Scheduler", "Logging", "Machine Learning", "MapReduce", "Miscellaneous", "Natural Language Processing", "Network Virtualization", "Networking", "News Feed", "ORM", "Package Management", "Package Repositories", "Permissions", "Processes", "Queue", "RESTful API", "RPC Servers", "Science", "Search", "Serialization", "Serverless Frameworks", "Specific Formats Processing", "Static Site Generator", "Tagging", "Template Engine", "Testing", "Text Processing", "Third-party APIs", "URL Manipulation", "Video", "Web Content Extracting", "Web Crawling", "Web Frameworks", "WebSocket", "WSGI Servers", "Services", "Code Quality", "Continuous Integration", "Resources", "Podcasts", "Twitter", "Websites", "Weekly", "Other Awesome Lists", "Contributing"]
