--- pip_gui/Resource_Files/genreTagFile.json
@@ -1,1 +1,0 @@
-["#configuration", "#static-site-generator", "#search", "#package-repositories", "#debugging-tools", "#database-drivers", "#third-party-apis", "#wsgi-servers", "#game-development", "#concurrency-and-parallelism", "#science", "#queue", "#cms", "#build-tools", "#other-awesome-lists", "#restful-api", "#websites", "#microsoft-windows", "#processes", "#text-processing", "#audio", "#job-scheduler", "#gui", "#admin-panels", "#hardware", "#caching", "#tagging", "#anti-spam", "#html-manipulation", "#resources", "#computer-vision", "#miscellaneous", "#template-engine", "#asset-management", "#logging", "#functional-programming", "#authentication", "#data-visualization", "#video", "#code-analysis-and-linter", "#news-feed", "#serialization", "#code-quality", "#date-and-time", "#foreign-function-interface", "#documentation", "#networking", "#chatops-tools", "#geolocation", "#permissions", "#podcasts", "#services", "#cryptography", "#weekly", "#internationalization", "#deep-learning", "#specific-formats-processing", "#command-line-tools", "#distribution", "#machine-learning", "#network-virtualization", "#testing", "#orm", "#web-content-extracting", "#rpc-servers", "#http", "#downloader", "#files", "#web-frameworks", "#environment-management", "#serverless-frameworks", "#implementations", "#package-management", "#natural-language-processing", "#data-analysis", "#e-commerce", "#web-crawling", "#imagery", "#contributing", "#algorithms-and-design-patterns", "#devops-tools", "#websocket", "#editor-plugins-and-ides", "#mapreduce", "#continuous-integration", "#forms", "#url-manipulation", "#email", "#twitter", "#interactive-interpreter", "#data-validation", "#database", "#compatibility"]
