--- pip_gui/Resource_Files/installedPackageList.json
@@ -1,1 +1,0 @@
-["AdvancedHTTPServer", "Automat", "BBQSQL", "BTrees", "BeautifulSoup", "BlindElephant", "BusScraper", "Cheetah", "ConfigArgParse", "Django", "EditorConfig", "Flask", "Flask-FlatPages", "FormEncode", "Frozen-Flask", "GeoIP", "GitPython", "IPy", "Jinja2", "Mako", "Markdown", "MarkupSafe", "NfSpy", "PAM", "Paste", "PasteDeploy", "PasteScript", "Pillow", "PyGithub", "PyJWT", "PyMouse", "PyPDF2", "PyQRCode", "PySocks", "PyX", "PyYAML", "Pygments", "SQLAlchemy", "SecretStorage", "Tempita", "Twisted", "Werkzeug", "Whoosh", "XlsxWriter", "adns-python", "alembic", "appdirs", "argcomplete", "argh", "armstrong", "asn1crypto", "attrs", "awscli", "awssamllinux", "backports-abc", "backports.shutil-get-terminal-size", "backports.ssl-match-hostname", "basemap", "beautifulsoup4", "blinker", "boltons", "boto", "botocore", "brotlipy", "bs4", "capstone", "cassandra-flask-sessions", "certifi", "cffi", "chardet", "cheap", "chirp", "click", "colorama", "configobj", "configparser", "constantly", "construct", "couchdbkit", "cssutils", "cycler", "dap", "decorator", "dicttoxml", "dissy", "distorm3", "dnspython", "docker-py", "docutils", "easygui", "ecdsa", "enum34", "et-xmlfile", "fabric", "feedparser", "flask", "funcsigs", "fuse-python", "future", "futures", "geoip2", "geojson", "gevent", "greenlet", "h2", "hpack", "html2text", "html5lib", "http-parser", "httplib2", "httpretty", "hyperframe", "icalendar", "idna", "impacket", "incremental", "ipaddress", "ipython", "ipython-genutils", "itsdangerous", "jdcal", "jmespath", "jsbeautifier", "jsonpickle", "jsonrpclib", "keepnote", "keyring", "keyrings.alt", "killerbee", "listas1aninhadas", "lxml", "maxminddb", "mechanize", "mercurial", "metaconfig", "mitmproxy", "mockito", "motor", "msgpack-python", "mysqlclient", "nassl", "netaddr", "nltk", "numpy", "olefile", "openpyxl", "opsutils", "ordereddict", "packaging", "paramiko", "passlib", "pathlib2", "pathtools", "pcapy", "pefile", "persistent", "pexpect", "pickleshare", "pluginbase", "ply", "prettytable", "prompt-toolkit", "psutil", "psycopg2", "ptyprocess", "py", "pyOpenSSL", "pyasn1", "pyasn1-modules", "pycparser", "pycrypto", "pycryptodome", "pycurl", "pydns", "pyenchant", "pygobject", "pyinotify", "pylibemu", "pymongo", "pymssql", "pyotp", "pyparsing", "pyperclip", "pyrit", "pyscard", "pyserial", "pysmi", "pysnmp", "pysnmp-apps", "pysnmp-mibs", "pysqlite", "pytest", "python-Levenshtein", "python-apt", "python-dateutil", "python-debian", "python-debianbts", "python-editor", "python-jose", "python-ntlm", "python-openid", "python-pam", "pytz", "pyusb", "pyxdg", "qrcode", "reportbug", "requests", "restkit", "rfidiot", "roman", "rsa", "s3transfer", "scandir", "scapy", "scgi", "search1337", "service-identity", "simple-crypt", "simplegeneric", "simplejson", "singledispatch", "six", "slowaes", "smoke-zephyr", "socketpool", "tblib", "tcpwatch", "termcolor", "torn", "tornado", "traitlets", "transaction", "typing", "tzlocal", "uTidylib", "urllib3", "urwid", "vinetto", "virtualenv", "volatility", "wafw00f", "wapiti", "watchdog", "wcwidth", "webencodings", "websocket-client", "webunit", "wfuzz", "wxPython", "wxPython-common", "xdot", "xmlbuilder", "yara-python", "zenmap", "zim", "zope.annotation", "zope.browser", "zope.component", "zope.configuration", "zope.container", "zope.contenttype", "zope.dottedname", "zope.event", "zope.exceptions", "zope.filerepresentation", "zope.i18n", "zope.i18nmessageid", "zope.interface", "zope.lifecycleevent", "zope.location", "zope.proxy", "zope.publisher", "zope.schema", "zope.security", "zope.site", "zope.size", "zope.traversing"]
