--- pip_gui/Resource_Files/installedPackageList3.json
@@ -1,1 +1,0 @@
-["Brlapi", "Django", "Jinja2", "Markdown", "MarkupSafe", "Pillow", "PyDictionary", "PyOpenGL", "PyYAML", "SecretStorage", "Werkzeug", "binwalk", "boto", "botornot", "chardet", "click", "crcelk", "cryptography", "cupshelpers", "cycler", "debtags", "decorator", "fabric", "flask", "flask-flatpages", "futures", "goslate", "hashID", "html5lib", "idna", "itsdangerous", "keyring", "keyrings.alt", "louis", "lxml", "matplotlib", "numpy", "oauthlib", "paramiko", "pyasn1", "pycrypto", "pycups", "pycurl", "pygobject", "pyparsing", "pyqtgraph", "pyserial", "pysmbc", "python-apt", "python-dateutil", "python-debian", "pytz", "pyxdg", "requests", "requests-oauthlib", "scipy", "search1337", "six", "smoke-zephyr", "termineter", "torn", "tornado", "tweepy", "unattended-upgrades", "urllib3", "webencodings"]
