--- pip_gui/Resource_Files/outdatedPackageList.json
@@ -1,1 +1,0 @@
-["docker-py", "jmespath", "keyring", "mercurial", "openpyxl", "paramiko", "persistent", "PyGithub", "PyJWT", "pylibemu", "pyOpenSSL", "pyscard", "pysmi", "python-dateutil", "pytz", "PyX", "requests", "s3transfer", "tornado", "transaction", "tzlocal", "webencodings"]
