--- pip_gui/Scraping/githubFile.py
@@ -1,2 +1,0 @@
-import os
-os.system('curl https://github.com/vinta/awesome-python/blob/master/README.md >> db.txt')
