+++ pipgui/Resource_Files/installedPackageList.json
@@ -1,0 +1,1 @@
+["1337", "astroid", "autopep8", "beautifulsoup4", "certifi", "chardet", "Click", "colorama", "docopt", "Flask", "idna", "isort", "itsdangerous", "Jinja2", "lazy-object-proxy", "lxml", "MarkupSafe", "mccabe", "oauthlib", "pipreqs", "pycodestyle", "pylint", "PyQt5", "PyQt5-sip", "PySocks", "requests", "requests-oauthlib", "six", "soupsieve", "typed-ast", "urllib3", "Werkzeug", "wrapt", "yarg"]
