+++ pipgui/Resource_Files/installedPackageList3.json
@@ -1,0 +1,1 @@
+["1337", "Click", "Flask", "Jinja2", "MarkupSafe", "PyQt5", "PyQt5-sip", "PySocks", "Werkzeug", "astroid", "autopep8", "beautifulsoup4", "certifi", "chardet", "colorama", "docopt", "idna", "isort", "itsdangerous", "lazy-object-proxy", "lxml", "mccabe", "oauthlib", "pipreqs", "pycodestyle", "pylint", "requests", "requests-oauthlib", "six", "soupsieve", "typed-ast", "urllib3", "wrapt", "yarg"]
