+++ pipgui/Scraping/githubFile.py
@@ -1,0 +1,2 @@
+import os
+os.system('curl https://github.com/vinta/awesome-python/blob/master/README.md >> db.txt')
