+++ pipgui/__init__.py
@@ -1,0 +1,2 @@
+
+from mainGUI import main
